pyproject.toml
[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[tool.uv.workspace]
members = ["packages/client", "packages/server"]

[tool.uv.sources]
memmachine-server = { workspace = true }
memmachine-client = { workspace = true }

[project]
name = "memmachine"
version = "0.1.0"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.31.0",
    "aiosqlite>=0.20.0",
    "boto3>=1.40.40",
    "boto3-stubs==1.40.64",
    "coverage>=7.11.0",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "fastmcp>=2.12.0",
    "instructor[bedrock]>=1.12.0",
    "langchain-aws>=0.2.33",
    "neo4j>=5.28.2",
    "nltk>=3.9.1",
    "openai>=1.104.2",
    "pgvector>=0.4.1",
    "prometheus-client>=0.22.1",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "rank-bm25>=0.2.2",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "greenlet>=3.2.4",
    "tzdata>=2024.1 ; sys_platform == 'win32'",
    "memmachine-server",
    "memmachine-client",
    "flagembedding>=1.3.5",
    "kneed>=0.8.5",
    "cvxpy>=1.7.5",
    "langchain-text-splitters>=1.1.0",
    "spacy>=3.8.11",
    "stanza>=1.11.0",
    "pip>=25.3",
    "google-genai>=1.57.0",
]
requires-python = ">= 3.12"
authors = [
  {name="MemMachine", email="noreply@memmachine.ai"},
]
description = "MemMachine is an open-source memory layer for AI agents, giving your creations short-term, long-term, and personalization memory, enabling them to learn, adapt, and remember."
readme = "README.md"
keywords = ["memmachine"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "Apache-2.0"
license-files = ["LICENSE"]

[project.optional-dependencies]
gpu = [
    "sentence-transformers>=5.1.0",
]

[project.scripts]
memmachine-server = "memmachine.server.app:main"
memmachine-sync-profile-schema = "memmachine.profile_memory.storage.syncschema:main"
memmachine-nltk-setup = "memmachine:setup_nltk"
memmachine-mcp-stdio = "memmachine.server.mcp_stdio:main"
memmachine-mcp-http = "memmachine.server.mcp_http:main"

[project.urls]
Homepage = "https://github.com/MemMachine/MemMachine"
Issues = "https://github.com/MemMachine/MemMachine/issues"

[dependency-groups]
dev = [
    "complexipy>=4.2.0",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.13.2",
    "testcontainers[neo4j,postgres]>=4.13.1",
    "ty>=0.0.1a26",
]

[tool.mypy]
follow_untyped_imports = true
enable_error_code = ["deprecated"]
exclude = ["venv/", ".venv/"]
plugins = ['pydantic.mypy']

[tool.ruff]
exclude = ["tools"]

[tool.ruff.lint]
select = [
    "A",
    "ANN",
    "ARG",
    "ASYNC",
    "B",
    "C4",
    "C90",
    "D",
    "DTZ",
    "E",
    "F",
    "FA",
    "FAST",
    "FURB",
    "G",
    "I",
    "ICN",
    "INP",
    "INT",
    "ISC",
    "LOG",
    "N",
    "NPY",
    "PERF",
    "PIE",
    "PLC",
    "PT",
    "PTH",
    "PYI",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "SLF",
    "SIM",
    "T20",
    "TID",
    "TRY",
    "UP",
    "W",
    "YTT",
]
ignore = [
    "E501", "UP015", "TRY002", "PLC0415",
    "D102", "D212", "D203", "RET504",
    "TRY003",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN", "D", "SLF", "ARG"]
"tools/**/*.py" = ["SIM", "DTZ", "C9", "ANN", "T201", "PT", "INP", "PTH", "D"]
"examples/**/*.py" = [
    "ANN",
    "ASYNC210",
    "T201",
    "D",
    "ARG",
    "T201",
    "TRY401",
    "C901",
    "TRY300",
    "PTH",
]
"evaluation/**/*.py" = [
    "D",
    "ARG",
    "T201",
    "ANN",
    "PTH123",
    "PTH120",
    "PTH110",
    "PTH100",
    "ASYNC230",
    "RUF001",
    "N806",
    "B023",
    "B007",
    "TRY300",
    "PERF401",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"

markers = [
    "integration: mark the test as an integration test",
    "slow: mark the test as slow"
]

[tool.complexipy]
paths = ["src", "tests"]
max-complexity-allowed = 10
quiet = false
ignore-complexity = false
details = "low"
color = "auto"
sort = "desc"
exclude = []

[tool.coverage.run]
source = [
  "src/memmachine",
]

[tool.coverage.report]
sort = "cover"
